+++ packages/usb_iso_cli/test/usb_iso_cli_test.dart
@@ -17,11 +17,19 @@ void main() {
 
   test('make help mentions repeating --iso for multiboot', () async {
     expect(await run(['make', '--help']), 0);
+    final make = MakeCommand();
+    expect(make.description, contains('multiboot'));
+    expect(make.argParser.options['iso']!.help, contains('Repeat'));
   });
 
-  test('add and refresh help succeed', () async {
+  test('add and refresh help describe the non-destructive path', () async {
     expect(await run(['add', '--help']), 0);
     expect(await run(['refresh', '--help']), 0);
+    final add = AddCommand();
+    expect(add.description, contains('without erasing'));
+    expect(add.argParser.options['yes']!.help, contains('ADD'));
+    final refresh = RefreshCommand();
+    expect(refresh.description, contains('GRUB'));
   });
 
   test('rewrites a TTY write line and prints other steps once', () {